Change "capacity" to "intrinsic_growth" for Kisdi.
The first term in the Kisdi model is not a carrying capacity (the
equilibrium population), it is the intrinsic growth rate in the absence
of competition (including self-competition), like $r$ in a classic
Lotka-Volterra model.

I updated the method and argument names accordingly.

I also added some units to parameter documentation.